% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOdedup.R
\name{merge.similar.terms}
\alias{merge.similar.terms}
\title{merge similar GO term for dataframe output from gprofiler2}
\usage{
\method{merge}{similar.terms}(gostres.tab)
}
\arguments{
\item{gostres.tab}{result table from gprofiler2::gost. intersection column is required in dataframe.}
}
\value{
a subset dataframe of input
}
\description{
merge similar GO term for dataframe output from gprofiler2
}
\details{
none
}
\author{
ShaoboLiang
}
